QGD-Net-main/requirements.txt
matplotlib
numpy
Pillow
tqdm
wandb
torch
